Refactored MultiStepForm: extracted utilities to formUtils.js
